crypto/keys_test.go
package crypto

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestGeneratePrivateKey(t *testing.T) {
	privKey := GeneratePrivateKey()
	assert.Equal(t, len(privKey.Bytes()), privKeyLen)
	pubKey := privKey.Public()
	assert.Equal(t, len(pubKey.Bytes()), pubKeyLen)
}

func TestPrivateKeySign(t *testing.T) {
	privKey := GeneratePrivateKey()
	pubKey := privKey.Public()
	msg := []byte("A message to be pass over the wire")
	sign := privKey.Sign(msg)

	assert.True(t, sign.Verify(pubKey, msg))
	assert.False(t, sign.Verify(pubKey, []byte("A different message")))

	// Test with Invalid Public Key
	invalidPrivKey := GeneratePrivateKey()
	invalidPubKey := invalidPrivKey.Public()
	assert.False(t, sign.Verify(invalidPubKey, msg))
}
